Implement static global aliases
The main reason why implementing global aliases is difficult is the
updating in case of added or removed global aliases, because a bot may
have locally overwritten a global alias, which of course should not be
removed if a global alias is removed.

This patch circumvents that by not updating global aliases at all.
They get set from the current list whenever a bot is started, but the
configuration of bots does not change when the global list is changed.